move the statistic calculation logic into a separate IMT
