+++ src/features/requestmonitor.feature
@@ -4,13 +4,6 @@ Feature: Request Monitor service
 
   Scenario: Register incoming request
     When the system get an Incoming request
-    Then I see "1" for "totalIncomingRequests" in the statistics
-    Then I see "1" request in the database
-
-  Scenario: Recalculate the requests
-    Given the system get an Incoming request
-    When the system recalculate the requests
-    Then I see "1" for "totalIncomingRequests" in the statistics
     Then I see "1" request in the database
