+++ src/features/steps/requestmonitor.js
@@ -8,21 +8,9 @@ export default function () {
     await monitor.registerIncomingRequest('url');
   });
 
-  this.When('the system recalculate the requests', async function () {
-    const monitor = this.container.get('requestmonitor');
-    await monitor.recalculate();
-  });
-
-  this.Then('I see "$value" for "$key" in the statistics', async function(value, key) {
-    const monitor = this.container.get('requestmonitor');
-    const result = await monitor.getStatistics();
-    expect(result[key]).to.eql(parseInt(value));
-  });
-
   this.Then('I see "$value" request in the database', async function(value) {
     const monitor = this.container.get('requestmonitor');
     const result = await monitor.getRequests();
     expect(result.length).to.eql(parseInt(value));
   });
-
 }
