+++ src/server/recalculate.js
@@ -1,9 +1,9 @@
 import container from './container';
 import VError from 'verror';
 
-const requestMonitor = container.get('requestmonitor');
+const stats = container.get('stats');
 
-requestMonitor.recalculate().then(()=>{
+stats.recalculate().then(()=>{
   console.log('recalculation was successful');
 }).catch((error)=>{
   console.log(new VError(error, `[Recalcuate] The recalcuation was unsuccessful`));
